refactor: extract UndoMixin, ContentMixin, RenderMixin from widget.py
Extract three new mixins to continue the widget.py decomposition:
- UndoMixin (action/undo.py): undo/redo and dot-repeat logic
- ContentMixin (action/content.py): content get/set, JSON validation,
  formatting, embedded JSON editing, and JSONL helpers
- RenderMixin (action/render.py): render(), syntax highlighting,
  gutter rendering, wildmenu, and display utilities

widget.py reduced from 1,276 to 367 lines (71% reduction).
All 402 tests pass with no changes to test code.
