+++ src/jvim/action/__init__.py
@@ -1,15 +1,21 @@
 """Action mixins and utilities for JsonEditor."""
 
 from jvim.action.clipboard import ClipboardMixin
+from jvim.action.content import ContentMixin
 from jvim.action.folding import FoldMixin
 from jvim.action.navigation import NavigationMixin
+from jvim.action.render import RenderMixin
 from jvim.action.substitute import SubstituteMixin
+from jvim.action.undo import UndoMixin
 from jvim.action.visual import VisualMixin
 
 __all__ = [
     "ClipboardMixin",
+    "ContentMixin",
     "FoldMixin",
     "NavigationMixin",
+    "RenderMixin",
     "SubstituteMixin",
+    "UndoMixin",
     "VisualMixin",
 ]
